requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
httpie==2.3.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
pycodestyle==2.6.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
PySocks==1.7.1
pytz==2020.4
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
